fix: Handle missing fields on TCO2 credit class page (#1817)
* fix: use custom startCase fn instead of lodash which rm special char

* refactor: mv getValue to utils

* fix: inconsistent regen:approvedMethodologies fields
