+++ web-registry/src/pages/CreditClassDetails/CreditClassDetails.AdditionalInfo.tsx
@@ -1,6 +1,5 @@
 import { Box, Grid } from '@mui/material';
 import { ClassInfo } from '@regen-network/api/lib/generated/regen/ecocredit/v1/query';
-import { startCase } from 'lodash';
 
 import { Body } from 'web-components/lib/components/typography';
 
@@ -10,6 +9,7 @@ import { ArrowLink } from 'components/atoms/MetadataArrowLink';
 import { MetaDetail } from 'components/molecules';
 
 import { ApprovedMethodologiesList } from './CreditClassDetails.ApprovedMethodologies';
+import { getValue } from './CreditClassDetails.utils';
 
 interface AdditionalInfoProps {
   onChainClass: ClassInfo;
@@ -26,13 +26,6 @@ const AdditionalInfo: React.FC<React.PropsWithChildren<AdditionalInfoProps>> =
     const projectActivities = metadata?.['regen:projectActivities'];
     const carbonOffsetStandard = metadata?.['regen:carbonOffsetStandard'];
     const tokenizationSource = metadata?.['regen:tokenizationSource'];
-
-    // This can be deleted if all class metadata is updated to latest standard
-    const getValue = (val: any): string => {
-      let value = val;
-      if (val?.['@value']) value = val['@value'];
-      return startCase(value);
-    };
 
     const getCreditType = (creditTypeAbbrev: string): string => {
       // TODO: add credit types as they come online, or fetch from ledger somehow
